FEATURE: added auto-db backups
This makes a copy of the entire sqlite database into a backups/ folder, datestamped in UTC

- This is controlled by a Preferences pane that allows you to say how many backups you want to keep, and how often they should be written.
- You can force a backup right now via Tools -> Backup Now
- You can restore a previous backup via Tools -> Restore
- You can look at all your backups via Tools -> Backups

Also, we now use the backup system to restore the database if we fail in our migrations at app startup on a new version of the app.  We restore the backup, display a stack trace, and tell the user which version they had previously used so they can downgrade to that version to continue working.